dashboard-superset/requirements.txt
# Dependências Python - Dashboard Superset

# Parsing e processamento
markdown==3.4.4
python-frontmatter==1.0.0

# Banco de dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
pandas==2.1.3
python-dotenv==1.0.0

